improve(mobile): move battery to user_screen
